+++ packages/client/src/components/admin/manage/cad-settings/MiscFeatures.tsx
@@ -6,7 +6,7 @@ import { FormField } from "components/form/FormField";
 import { Loader } from "components/Loader";
 import { useAuth } from "context/AuthContext";
 import useFetch from "lib/useFetch";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { FormRow } from "components/form/FormRow";
 import { MiscCadSettings } from "types/prisma";
