+++ packages/client/src/components/admin/manage/users/BanArea.tsx
@@ -1,6 +1,6 @@
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Loader } from "components/Loader";
 import { useAuth } from "context/AuthContext";
 import { Formik } from "formik";
