+++ packages/client/src/components/admin/values/import/ImportValuesModal.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Loader } from "components/Loader";
 import { Modal } from "components/modal/Modal";
 import { Formik, FormikHelpers } from "formik";
