+++ packages/client/src/components/admin/values/penal-codes/ManagePenalCode.tsx
@@ -1,7 +1,7 @@
 import { CREATE_PENAL_CODE_SCHEMA } from "@snailycad/schemas";
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Loader } from "components/Loader";
 import { Modal } from "components/modal/Modal";
 import { Formik, useFormikContext } from "formik";
