+++ packages/client/src/components/admin/values/penal-codes/ManagePenalCodeGroup.tsx
@@ -1,6 +1,6 @@
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Loader } from "components/Loader";
 import { Modal } from "components/modal/Modal";
 import { useModal } from "context/ModalContext";
