+++ packages/client/src/components/bleeter/ManageBleetModal.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/router";
 
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Loader } from "components/Loader";
 import { Modal } from "components/modal/Modal";
 import { useModal } from "context/ModalContext";
