+++ packages/client/src/components/ems-fd/modals/CreateMedicalRecord.tsx
@@ -10,7 +10,7 @@ import { useModal } from "context/ModalContext";
 import { ModalIds } from "types/ModalIds";
 import type { MedicalRecord } from "types/prisma";
 import { handleValidate } from "lib/handleValidate";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Textarea } from "components/form/Textarea";
 import { Select } from "components/form/Select";
 import { useCitizen } from "context/CitizenContext";
