+++ packages/client/src/components/leo/modals/CreateRecord/TableItemForm.tsx
@@ -1,5 +1,5 @@
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { SelectValue } from "components/form/Select";
 import { useFormikContext } from "formik";
 import { useTranslations } from "next-intl";
