+++ packages/client/src/components/leo/modals/WeaponSearchModal.tsx
@@ -8,7 +8,7 @@ import { Form, Formik } from "formik";
 import useFetch from "lib/useFetch";
 import { ModalIds } from "types/ModalIds";
 import { useTranslations } from "use-intl";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Citizen, Value, Weapon } from "types/prisma";
 import { Infofield } from "components/shared/Infofield";
