+++ packages/client/src/components/truck-logs/ManageTruckLog.tsx
@@ -2,7 +2,7 @@ import { CREATE_TRUCK_LOG_SCHEMA } from "@snailycad/schemas";
 import { Button } from "components/Button";
 import { FormField } from "components/form/FormField";
 import { FormRow } from "components/form/FormRow";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Select } from "components/form/Select";
 import { Loader } from "components/Loader";
 import { Modal } from "components/modal/Modal";
