+++ packages/client/src/pages/admin/manage/users/[id].tsx
@@ -19,7 +19,7 @@ import { Toggle } from "components/form/Toggle";
 import { FormRow } from "components/form/FormRow";
 import { BanArea } from "components/admin/manage/BanArea";
 import { handleValidate } from "lib/handleValidate";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { requestAll } from "lib/utils";
 import { DangerZone } from "components/admin/manage/users/DangerZone";
 import { Title } from "components/shared/Title";
