+++ packages/client/src/pages/admin/values/driverslicense-category.tsx
@@ -11,7 +11,7 @@ import useFetch from "lib/useFetch";
 import { Loader } from "components/Loader";
 import { AdminLayout } from "components/admin/AdminLayout";
 import { requestAll } from "lib/utils";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { FormField } from "components/form/FormField";
 import dynamic from "next/dynamic";
 import { SortableList } from "components/admin/values/SortableList";
