+++ packages/client/src/pages/admin/values/penal-code.tsx
@@ -14,7 +14,7 @@ import { requestAll } from "lib/utils";
 import dynamic from "next/dynamic";
 import { Table } from "components/shared/Table";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { ArrowLeft } from "react-bootstrap-icons";
 import { ModalIds } from "types/ModalIds";
 import { ManagePenalCodeGroup } from "components/admin/values/penal-codes/ManagePenalCodeGroup";
