+++ packages/client/src/pages/citizen/[id]/edit.tsx
@@ -8,7 +8,7 @@ import Link from "next/link";
 import { Button } from "components/Button";
 import { FormRow } from "components/form/FormRow";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import { Layout } from "components/Layout";
 import { Loader } from "components/Loader";
 import { handleValidate } from "lib/handleValidate";
@@ -21,7 +21,7 @@ import { useCitizen } from "context/CitizenContext";
 import { useValues } from "context/ValuesContext";
 import { requestAll } from "lib/utils";
 import { useAuth } from "context/AuthContext";
-import { ImageSelectInput, validateFile } from "components/form/ImageSelectInput";
+import { ImageSelectInput, validateFile } from "components/form/inputs/ImageSelectInput";
 import { Title } from "components/shared/Title";
 
 export default function EditCitizen() {
