+++ packages/client/src/pages/officer/supervisor/citizen-logs.tsx
@@ -9,7 +9,7 @@ import { GetServerSideProps } from "next";
 import { Citizen, RecordLog, RecordType } from "types/prisma";
 import { Table } from "components/shared/Table";
 import { FormField } from "components/form/FormField";
-import { Input } from "components/form/Input";
+import { Input } from "components/form/inputs/Input";
 import format from "date-fns/format";
 import { useGenerateCallsign } from "hooks/useGenerateCallsign";
 import { ArrowLeft } from "react-bootstrap-icons";
